feat: Move component dependencies
